[#171433917] Downgrade webpack-dev-server, it was broken
There is too many changes, this was previously upgraded
to get security patches. However in version 3 they split out webpack-cli
which breaks it from being able to run, and then made a lot of other
incompability changes, which probably requires
webpack upgrades, which requires a rewrite of the webpack.config.js and
so on; never ending process...

This at least upgrades to the latest 2.x version.