[#169176122] Update links to open in new window
The goal of this story is to update internal links to external sites
to use `_blank`, opening the site in a new window, with `noopener
norefferer` to prevent hijacking of the origin tab (Promote in this
case).

This commit progresses this story by updating Quill links to have
these attributes.